refactor: Rename package name from pieces to piece
Renamed package name from pieces to piece.
Added private constructor in PrintUtils to prevent from being instance.
