+++ src/main/java/piece/Position.java
@@ -1,4 +1,4 @@
-package pieces;
+package piece;
 
 import java.util.Arrays;
 import java.util.List;
